video-compressor/requirements.txt
python-dotenv
kafka-python-ng
ffmpeg-python
minio